increase the tt.value from short to int.  This holds score and it was overflowing.
TT entry size should go from 16 to 18 bytes.

+++ sillychess/src/sillychess.h
@@ -120,7 +120,7 @@ typedef struct tagHASHE {
     u64 key;
     unsigned char depth;
     unsigned char flags;    // bits 0,1,2,3 store the type (exact,alpha,beta etc)
-    short value;
+    int value;
     unsigned int bestMove;
 } Hash_Entry;
